requirements.txt
asgiref==3.4.1
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-cors-headers==3.7.0
mysqlclient==2.0.3
gunicorn==20.1.0
cryptography==3.4.8
channels==3.0.4
channels_redis==3.4.0
Pillow==9.3.0